Adds "expression too complex" bail reasons
Instead of just returning a bool, the reason the solver ends early is encoded
into an enum, which is then bubbled up to separate error messages.
To decrease review cost of this PR, a wrapper call maintains the original bool
semantics for client code.

Multiple error messages have been added. They each contain the same core
explanation of "the expression is too complex". However, they now are broken out
per termination reason.

While this itself is not an enormous improvement for end-users it does
allow people googling specific problems to have slightly more to go on
for which reason bumped them out of the solver run.

This is also scaffolding for future work. That can bubble up
useful info to the user such as the type assignments which have
been made for all disjunctions, and which have not. This could also
lead to showning which sub-expressions currently have many disjunctions
(and which don't). If nothing else, that would give users a start on
*a* simplification that would get their code typechecking. Something like:

Line 56: The expression is inferring 12 types. Please rewrite the expression
in more lines with more explicit types. \<click to expand for example\>.

Even imperfect information here can be a huge help to those who are
used to looser conversion semantics of other languages and to langeuages
where you always write types everywhere.