Get .mean() and similar ops to work without host copies in simple cases when the result is not used in the tensor program:  - Introduce _TFGetScalarOrDie for ops that known to produce 0d tensors, so the    compiler doesn't think that these are failable conditions.  - Introduce indirection through AccelerableTensorUnit to implement this, which    avoid generics introducing address-only out parameters to the low level    primitives that partitioning wants to work on.  - Teach canMoveInstruction that it can hoist or sink __tf_get_scalar_or_die    primitives now that they don't take things by-address.  - Switch mean/min/max/sum to use the new thing.
Change the newly introduced varargs mean(alongAxes: Int...) function to take
its axes as an array instead of a varargs list.  This sidesteps the recent
controversy with empty varargs, but more importantly, fixes an apparent
ambiguity where the compiler was picking this function instead of the zero
argument one which is more specific.  This seems like a swift bug.

Taking an array for this is not ideal, but it seems actually sort of reasonable
if we introduce mean(axis: Int, keepingDimensions: Bool = false) to go along
with it.  Something to discuss in more detail tomorrow.

PiperOrigin-RevId: 184843484