Fix `transposed`, remove `reversed`. Disable GPU tests involving loops.
The key problem with `reversed` is ReverseV2 does not support Int64 on GPU.
`transposed` has been updated to not use `reversed` when calculating default permutations.
`reversed` has been removed because it's purpose for inclusion was to support `transposed`.

The ideal solution to the problem is add Int64 support to ReverseV2 on GPU.
Removing Int from AccelerableTensorUnit is not a permanent solution because the problem is more broad: what's really needed is a way to limit ops to a subset of datatypes. This is currently doable but ugly and not generalizable.

GPU tests involving tests are disabled.

PiperOrigin-RevId: 184917302