[silgen] All ownership endowed APIs on SILGenBuilder should use SILGenBuilder SILValue APIs to ensure we get proper conformances.
I was thinking about this last night. I think the reason that this difference
came up is that people saw the SILGenBuilder SILValue based conformance APIs and
copied their calling of SILBuilder SILValue APIs. In the general case this is
not safe, since in some cases we need to add conformances.

Noticed by inspection.