[+0-normal-args] When initializating an Any with trivial inline buffer contents, put a cleanup on the Any.
Otherwise, it looks like we are storing a non-trivial value (Any) to memory
without a +1 cleanup, violating invariants. The change itself is harmless since
destroying an Any containing only trivial things in the inline buffer is a
no-op.

This is the last change I need to enable the only +1 into memory assert.

rdar://34222540