[+0-normal-args] Fix a typo in createTuple where we were looking for the first trivial element instead of the first non-trivial element.
This would cause us to not properly create cleanups on tuples, causing potential
memory leaks. Luckily, I verified that we previously did not call this anywhere
in the compiler by removing the method in question from SILGenBuilder and
performing a successful build of swift. So we do not need to worry about this
bug on 4.1/5.0. I think I added it just for completeness since we had other such
ManagedValue methods.

Caught by the sil ownership verifier while running the argument shuffle test
with +0 rvalues and the commit merged with this commit.

rdar://34222540