[+0-normal-args] Be sure if we are forwarding arguments into memory, that we do it at +1.
This can only happen when +0 is enabled. The reason why is that guaranteed self
is immutable so can not be passed inout without going through a var. There are
no other cases I can think of where an ArgumentSource will put self into
memory.

This is tested by the updated SILGen +0 tests.

rdar://34222540