[lldb-with-tools] Add a new utility script called lldb-with-tools.
This is a simple tool that starts lldb, but before it runs your commands, uses
the -O command to load lldbToolBox.py. This provide sthe llvm data formatters as
well as potential future lldb extensions that we write for swift itself.

A better name for this utility would be much appreciated, but I think for now
the name "lldb-with-tools" is at least self explanatory.