Fix Tensor/TensorXD test breakage caused by incorrect use of self.init().
self.init() shouldn't be called because it's calling the variadic initializer which then calls the caller of self.init().

Fixed the problem by adding @_fixed_layout to TensorXD and TensorShape. Requiring @_fixed_layout may be either a DI bug or a temporary requirement for upcoming changes in the Swift development branch.

PiperOrigin-RevId: 186160882