[lldb-tools] Add command disassemble-asm-cfg.
This works by dumping the disassembly of the current assembly frame into a
temporary file and then running:

  blockifyasm < tmpfile | viewcfg

without having to leave lldb.

I also fixed a small bug in the lldb-with-tools.in script where the input was
passed onto lldb as:

  lldb -O '...' -- $@

The -- should not have been there since sometimes one wants to /not/ use a --
argument form to lldb.