StdlibUnittest: zero out the hash key in TestSuite.init.
This eliminates the chance of random test failures due to changing hash values, freak hash collision situations and unstable Dictionary ordering.

On the other hand, it allows tests to rely on specific hash values, which is unwise. It also requires that a TestSuite be initialized before the first element is inserted into any Set or Dictionary.